Added the Random DDGs folder.
The new folder has a .tgffopt file for generating random graphs, and a .l file to parse the tgff output into a file that can be fed to the python script. The FHE model with limited bootstrapping seems to be working, but testing with random graphs will give a better idea.