Added make_graph.sh script and validate_unlimited_ilp_solution.py.
Random DDGs folder has been renamed to DDGs. It contains a bash script called make_graph.sh that accepts one argument, the desired name of the output files, to which the appropriate extenstions will be appended. It uses TGFF to make the random graphs, then runs a lex program and a python script to convert TGFF's output into something that can be passed to Lingo. Some random graphs have been generated and tested on the unlimited and bootstrap-limited FHE models. Only smaller graphs have solved in short amounts of time, at least on the bootstrap-limited version. The validate_unlimited_ilp_solution.py script is meant to ensure the solution given by the unlimited Lingo model actually follow constraints. It does not verify that it is an optimal solution.