Implemented list scheduling for the unlimited model.
Moved some functions from SolutionValidator class to the LGRParser class that are utilized by both the list scheduler and the solution validator. This reduces code repetition and LGRParser seemed to be the appropriate place for these functions because they only involve information found in lgr files.
Essentially completely rewrote list_scheduling.cpp to make it work in this project as opposed to its original use in a Compiler Design course. This includes changing it to a class, which involved creating a list_scheduling.h file. The input of the class is an FHE DAG and a minimum-bootstrapping lgr file. The output of the class is an lgr-like file that can be passed to the solution validator.
The SolutionValidator constructor no longer accepts a config struct, but rather two strings that are input file paths.