Added some shared_utils functions.
Moved remove_chars_from_string and extract_number_from_string functions from the LGRParser class to shared_utils, since they are generic functions, not specific to parsing lgr files. I added to the LGRParser class the add_info_to_operation_list function, since it is more appropriate to keep that functionality in that class, as opposed to writing it directly in the SolutionValidator constructor, where is previously was. Similarly, the function may be reused by the ListScheduler class in the future, so that will improve code reuse.