Switched from int IDs to shared_ptrs for accessing shared Operation objects.
Instead of having a vector of Operation objects that can be accessed via the operation's ID, Operation objects are stored as shared_ptrs. This allows the many data structures that stored operation IDs to directly store a shared_ptr to the Operation instead. Because many files had data structures storing IDs, those all needed to be updated.
Also, some code was moved from the InputParser class to the new BootstrappingPathGenerator class. This was done because creating bootstrapping paths is not actually part of parsing the DAG files. Rather it depends on information from the parsed DAG file. Furthermore, sometimes information from the LGRParser is required to generate bootstrapping paths, specifically whether the selective model was used. Since the LGRParser also requires information from the InputParser, it makes sense to separate BootstrappingPathGenerator into its own class so that the program can follow the general order of using these classes in a necessary order: InputParser -> LGRParser (if there is an lgr file to parse) -> BootstrappingPathGenerator.