Another heuristic option added for list scheduling. LimitedToSelectiveConverter class added.
Removed bug from GraphGenerator class that occurred when generating graphs with only one starting operation.
A new member variable has been added to the Operation struct. It is a vector of OperationPtrs. These operations are the children to which the operation sends bootstrapped results. This means that which operations are bootstrapped does not need to be kept in a separate data structure, and it reduces code differences when dealing with the limited model vs the selctive model.
A new class LimitedToSelectiveConverter takes a standard minimum bootstrapping lgr file as an input. It then converts this to a selective minimum bootstrapping lgr file by removing unnecessary bootstrapping operations in a selective context. Its usefulness has not been tested. However, if working properly, it cannot produce results worse than the limited list scheduling. Since the selective minimum bootstrapping lgr files have produced poor results when used for list scheduling, this new method should largely been an improvement.
The ListScheduler class now includes the option to perform list scheduling without predetermined operations to bootstrap. This is done by using a heuristic based on the number of bootstrapping paths an operation is included in, the rank (latest_start_time - earliest_start_time) of the operation, and how many children an operation has. The rank of all functions is recalculated every time another operation is chosen for bootstrapping. So far testing suggests this method is worse than minimum bootstrapping.
Many new test graphs have been generated using the custom random graph generator.