Fixed major LimitedToSelectiveConverter bug.
The LimitedToSelectiveConverter class had a bug that improperly handled removing an element from a vector while that vector was being iterated over. Fixing this solved an issue that caused necessary bootstrappings to be removed.
The bash script convert_multiple_limited_to_selective.sh was created and used to perform the conversion on all 100 test graphs. As expected, this leads to results that are either the same or better than the non-selective list scheduling version.