New bootstrapping selection heuristic based on urgency.
The ListScheduler class has been changed to use a heuristic based on the urgency of bootstrapping. The original heuristic code is commented out in the class. Testing so far shows the new heurisitic performs better for limited 1 core, but is not improved by limited-to-selective conversion, making it worse than the original heuristic for selective 1 core.

Files for using simulated annealing and a genetic algorithm to make schedules have been created, but almost no code has been written yet.