Updated Makefile for faster compilation, fixed path generator errors, fixed heuristic options, began changing bootstrapping urgency calculation.
Makefile in CPP_Code folder updated to make .o files, then the executables from those .o files. This should reduce compilation time, especially after editing only of the classes on which an executable depends.

Fixed two major errors in the BootstrappingPathGenerator class: One that excluded paths where the first operation has no parents, and one that read from the saved file incorrectly because ids were directly used to access the operations vector, instead of cllaing the get_operation_ptr_from_id function.

The heursitic in the ListScheduler class that was supposed to be "num_paths minus slack" was actually "num_paths minus urgency". Now the former is available as heuristic 3, and the latter is available as heuristic 4.

Work on a new method of calculating an operation's "bootstrapping urgency" has been started, but is not yet perfect. In particular, the criteria that determines if a path is "urgent" still needs some work.