Added path_nums to Operation struct, updated urgency criteria.
Added a path_nums variable to the Operation struct that tracks the indices of paths in the bootstrapping_paths vector that contain that operation, This is useful for quickly determining if an operation belongs to any bootstrapping paths, and may be useful in other ways. This information is added to the operations when the bootstrapping_paths vector is generated or read in from a file.

This was used to update the criteria for a path to be considered urgent. The first test showed that this criteria did not lead to a dead end.