Bootstrapping paths are now one operation longer.
Fixed an error with bootstrapping path generation that caused paths to be one operation shorter than they should be. The logic is used to find the paths, but they are then extended by one (adding all children of the last operation). This is necessary due to the difference between "selective" and standard paths.