Further improved efficiency of ListScheduler class.
The create_schedule operation has been removed. It was generally redundant, but some of the ideas used in it were migrated to the generate_start_times_and_solver_latency and generate_core_assignment functions. In particular, the pred_count data structure was added to these functions to greatly improve the efficiency of determining when an operation is ready.