Updated RandomGraphGenerator.
The format of DAG .txt files was updated while creating the benchmarks, but the RandomGraphGenerator (previously called GraphGenerator) was never updated to match it until now. Issues with the original graph generation were fixed too, such as the possibility of two redundant operations to exist (same operation e.g. ADD with the same inputs). The primary component of the update was accounting for "constant parents," so there are now some inputs related to how constant parents are added to operations. Also, the generation of a .vcg file was removed from the class, since the txt_to_vcg.cpp file can handle that.