Updated BootstrappingPathGenerator by assuming addition noise growth negligible.
Some of my recent testing seemed to suggest that, at least once noise is about where it would be after bootstrapping, the noise growth caused by addition and subtraction is negligible. Therefore, what constitutes a bootstrapping path could be slightly altered. In particular, paths should only end with an operation that has at least one child operation that is multiplication. Furthermore, it becomes clear that bootstrapping paths can only start with multiplication operations, though this was essentially already the case due to the requirement that the path's noise be under the threshold if the first operation on the path is removed. With this logic, it becomes unnecesary to track the number of additions along the path, and we simply need to compare the number of multiplications along the path to the number of levels gained via bootstrapping to know when to terminate a path.

The code in the BootstrappingPathGenerator class has been updated to reflect this, particularly the create_raw_bootstrapping_paths and create_bootstrapping_paths_helper functions. Furthermore, it reduces the number of constants that need to be changed for different numbers of bits from 4 constants to 1.