Added validation code to execution engine.
Added code functions to the execution engine that perform the FHE program on plaintext as well so the final results can be compared to the decrypted results of the actual FHE program. Any significant difference between the values (exact equivalence is not expected due to the approximate nature of CKKS) is reported.

Also, removed subtraction from RandomGraphGenerator for now. This avoids having subtractions with only one input (always resulting in zero).

Some small updates to scripts.