Added thread synchronization to execution engine.
The execution engine now has mutexes that can be used to create a post/wait-like mechanism. Only ciphertexts that will be written to get an associated mutex. The mutexes are locked on creation. When the ciphertext is written to, the mutex is unlocked. Any operations that require one of these ciphertexts as input must lock (and immediately) the mutex before performing the operation. Testing with random graphs confirms this solves the issues with multi-thread schedules that were not present in single-thread schedules. This synchronization mechanism only works properly because the schedules I have tested are in SSA form. Therefore, I will likely include code in a future commit that enforces SSA form.