Added basic bootstrapping mode to execution engine.
The execution engine now has a mode that will ignore the bootstrapping operations present in the input schedule and instead bootstrap ciphertexts as late as possible. After every operation, the output ciphertext's level is checked. If it is a certain value, that ciphertext is later used as an input, and this new mode is enabled, it will be bootstrapped and saved to the same variable. That is the primary change in the actual execution_engine.cpp file, while most of the changes are in the parser.cpp/hpp files, explained below.

With the new mode enabled, the parser does not add bootstrapping operations to the schedule, but rather keeps track of any bootstrapping operations' inputs and outputs. Then, any operations that had the output of a bootstrapping operation as an input, will have that input replaced with the input of the now-ignored bootstrapping operation. Furthermore, the parser keeps track of the set of ciphertexts that will be an input at some point, so that the execution engine can know if an operation needs to be bootstrapped.

Currently the variable that determines if this new mode is used is a constant defined at compile time, but will be updated to be a command line argument.