gained_levels now a command line argument.
Previously, gained_levels was a constant defined in shared_utils.h. But this sometimes led to mistakes where the gained_levels in that file was not the intended one and mismatched with the gained_levels assumed by a script. So it is now a command line argument to all C++ programs that utilize the BootstrappingPathGenerator class.