Major changes to directory system.
The layout of the DAGs and results directories have been updated to improve clarity. These directories are ignored by git, so the changes are primarily seen in scripts.

Also, fixed a bug in parser.cpp that caused initial_inputs to be empty in SCHED mode.

Added a test mode to the execution engine that only runs the schedule in plaintext, reporting any final values that are too small or too large. The execution engine now also writes the evaluation time and number of bootstrapping operations to files for later data collection.

Added a script that runs the same schedule (min_bootstrapping, certain heuristic, etc.) for multiple graphs.