Dead ciphertexts now deleted in execution engine.
Some FHE programs used a very large amount of memory because a lot of ciphertexts were being saved in the enc_regs map. However, many of those ciphertexts were not going to be used again (dead). To reduce memory usage in such cases, dead ciphertexts are erased from enc_regs. To achieve this, the outputs dependent on a certain ciphertext are tracked while parsing the schedule. Once all of these outputs have been calculated, that ciphertext is dead, so it is removed from enc_regs. Another set of mutexes is created to do this safely during multi-threaded execution.