Added Program, BootstrapSegment, and Operation classes.
A major refactor of the scheduling-related C++ code. Analysis of shared_utils showed that a number of functions would make more sense inside classes. For example, many functions took an Operation as an argument, but it would make more sense to call that operation as a member function of an Operation class. Therefore, the Program, BootstrapSegment, and Operation have been created and much code from shared_utils and other files has been moved to those classes. Currently, this code would compile if certain functions defined in .h files were implemented in the corresponding .cpp files. However, this commit is being made at this point because some of those classes/functions will be refactored instead of being implemented.

Also, the concept has been introduced that latency information will be stored in a file separate from the DAG description. This is because latency information is platform-specific, but the FHE program / DAG itself is not.