Changed Operation class back to a struct.
Operation has been made a struct again because it was cleaner for most of its member variables to be public, since a more class-like structure led to an excessive number of getters/setters. This new approach leaves a limited number of functions.

Program is now the only class that directly uses InputParser. This forces other files to use a Program class in place of their own separate OpVector master list of Operations. This is beneficial, since it gives access to the functions of the Program class, such as get_operation_ptr_from_id.

A number of other changes have been made so that compilation completes without errors. However, this code has not yet been tested.