Added reusable profiling function and improved segment generation speed.
Added some template classes and functions to shared_utils.h that offer a simple method of performing a function and printing the execution time to the console. To use this function, the programmer must create a std::function<T()> that runs the desired code to profile. That function, and a short string describing what the function does are the two inputs to the perform_func_and_print_execution_time function.

Using that profiling tool, it was discovered that segment generation was mainly slowed by two things: writing to file and the removal of redundant segments. The cost of removal of redundant segments has been significantly decreased by initially creating the standard segments (instead of selective segments), and removing redundant segments only at that point. Later, selective segments can be quickly generated from the standard without any redundant segments. This is a significantly faster process than the old one, which initially created selective segments, removed redundants, then converted to standard and removed redundants again. A significant source of the speedup is that there are never duplicate segments, so they do not need to be considered in the redundant removal process.

A future commit may restore the option to generate segments "on the the fly" for the C++ programs that require them, since that is seemingly faster than writing to / reading from a file. Alternatively reading/writing speeds may be improved.