Segments text file option reinstated.
Bootstrap segments can once again be saved as human-readable text. The default behavior is to create both binary and text files, but text files can be disabled to save time and space. However, there is not much time overhead, since the much faster writing method that is used for the binary files is now also used for the text files.