FileParser now a private class of the Program class.
Also, Program is now a friend class of FileParser. This allows the Program constructor to pass to FileParser a reference to the Program under construction. This way, FileParser can modify the Program directly instead of generating a new one and returning that. Also, instead of FileParser storing a shared_ptr to the Program, a reference_wrapper to the Program is stored, so it is clear that the FileParser does not share ownership of the Program. Similar changes have been made to LGRParser. Ideally, LGRParser would be a private class of FileParser, but since it relies on flexc++, that has not been implemented yet.