Created FileWriter class.
FileWriter is a private class of the Program class. The functions previously in the Program class that wrote to files are now in the FileWriter class. Instances of the Program class include an public member instance of the FileWriter class, which allows for the public file writing functions to be publicly accessed by Program instances, despite FileWriter being a private class.