Fixed segment_indexes variable of Operations.
At some point, the function that populated the segment_indexes member of Operations was commented out in the Program class. Performing that function is necessary for the calculation of bootstrap urgency to work as expected. The function has been brought back and added to the FileParser class. It is run at the end of parsing a segments file.