Program allows segments to be set, so BootstrapSegmentGenerator is working again.
Furthermore, convert_segments_to_selective is now a function of Program instead of BootstrapSegmentGenerator.

Other files that previously used the file_writer member of Program have been updated to make their own FileWriter instance and pass their Program to it. This includes CompleteToSelectiveConverter, which is no longer a class, but rather is implemented as a single main function in the file complete_to_selective_converter.cpp (the associated .h file has been deleted).

FileWriter is no longer a private class of Program, but is still a friend class.