Changed segment_indexes to unordered_set.
This further improves runtime of complete-to-selective conversion by allowing us to quickly check if a segment even contains the operations in a bootstrap pair before checking if that segment relies on the pair. However, parsing the segments file takes longer, since the code to populate the segment_indexes data structures is in that function.