Simplified calculation of earliest_start_time and latest_start_time.
The introduction of a public member variable "earliest_finish_time" helped with this. Furthermore, multiple bugs were fixed: earliest_start_time is set to zero before calculating its actual value, and the correct formula is now used for latest_start_time. Previously, these bugs caused the output of creating a bootstrap set to be nondeterministic and incorrect when including slack as a heuristic.