Improved runtime of calculating bootstrap_urgency via tracking alive segments.
By keeping track of which segments are alive, instead of checking every segment for its alive status, updating the bootstrap_urgency of every operation can be done much quicker. Updating which segments are alive is done more efficiently by using the information: which segments are newly satisfied, which operation was just chosen for bootstrapping, and which segments start with any given operation.