Execution engine can now read values from a file to use as the inputs to the program.
The execution engine now has three "input modes." RANDOM gives each input a random value between 0 and a given value. RANDOM is the default, with a value of 1.0.
CONSTANT gives all the inputs the same value, as defined on the commandline.
FILE reads a file and assigns each input a value based on that. The file is checked to have included values for all the inputs from the schedule, with no extra or missing inputs.

Also, added a commented out line to CMakeLists.txt that, if uncommented, compiles with debugging flags.