Changed how bootstrap segments are generated.
First, the method for finding ignorable operations is slightly changed. This means the generated segments will be different when initial_levels > 0. The seeds for random_graph1 and random_graph5 have been updated because this change made it so the previous seeds created graphs with no bootstrap segments when initial_levels=21.

Second, the segment generation process is now slower, because more segments are considered necessary than previously thought, specifically for selective bootstrap segments. So, the standard bootstrap segments that are generated are the same as before, but more selective bootstrap segments will be generated.