Fixed multi-threading bugs in execution engine.
Bugs occassionally occurred in the execution engine due to multiple threads accessing certain maps simultaneously. A simple mutex guarding access to the map solves these issues.

Also, a bug caused the order of some subtractions operands to be swapped when creating a schedule. This has been fixed by tracking the original string of operands in the Operation struct.